# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "SNS Application API",
    "description" : "社内向けSNSアプリケーションのREST API。\n\n## レスポンスの形\n\n成功時は `data` で包まれる。エラー時は `error` に `code` と `message` が入る。\n`code` はクライアントの分岐用、`message` はそのまま画面に出せる日本語。\n\n```json\n{ \"data\": { } }\n{ \"error\": { \"code\": \"POST_NOT_FOUND\", \"message\": \"投稿が見つかりません\" } }\n```\n\n## 認証（Swagger UI で試す手順）\n\n認証はhttpOnlyクッキー `auth_token` で行う。httpOnlyのためJavaScriptから読めず、\n**画面右上の Authorize ボタンは使えない**。代わりに次の手順で試す。\n\n1. `POST /api/auth/signup` または `POST /api/auth/login` を Try it out で実行する\n2. ブラウザがクッキーを保存する（このページはAPIと同じ `localhost:8080` から配信されているため）\n3. 以降のTry it outには自動でクッキーが乗る。Authorizeの操作は不要\n\nアクセストークンは15分で失効する。失効後は `POST /api/auth/refresh` を1回叩けば\n再発行される（リフレッシュトークンのクッキーは `Path=/api/auth` に限定されている）。\n\n## ページネーション\n\n一覧系はカーソルベース。`cursor` に前回応答の `nextCursor` を渡す。\n`nextCursor` が `null` なら末尾に到達している。\n\n## 画像\n\nレスポンスに含まれる画像URLはすべて**署名付きURLで有効期限は24時間**。\nURLそのものを永続化・長期キャッシュしないこと。\n\n## 関連ドキュメント\n\n設計の背景・採用理由・v1での割り切りは `docs/api-design.md` を参照。\nこちらは「何を・どんな型で」、api-design.mdは「なぜ」を扱う。\n",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "ローカル開発環境"
  } ],
  "security" : [ {
    "cookieAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/auth/signup" : {
      "post" : {
        "tags" : [ "認証" ],
        "summary" : "新規登録",
        "description" : "ユーザーを登録し、**同時にログイン状態にする**。\n成功時に `auth_token` と `refresh_token` の両クッキーがSet-Cookieされるため、\nクライアントは登録後にログインを呼び直す必要がない。\n\n成功時のステータスは **201 Created**。\n\nエラー: メールアドレスが登録済みの場合は400 `EMAIL_ALREADY_EXISTS`。\n",
        "operationId" : "signup",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SignupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "登録成功。認証クッキーが発行される",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "リクエストが不正（バリデーション違反、型不一致、壊れたJSON）",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "VALIDATION_ERROR",
                    "message" : "本文は280文字以内で入力してください"
                  }
                }
              }
            }
          },
          "429" : {
            "description" : "リクエストが多すぎる。Retry-Afterヘッダに再試行までの秒数が入る",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "RATE_LIMITED",
                    "message" : "アクセスが多すぎます。しばらく待ってからやり直してください"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "サーバー側の想定外エラー。詳細はクライアントへ返さずサーバーログに記録される",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "INTERNAL_ERROR",
                    "message" : "予期しないエラーが発生しました"
                  }
                }
              }
            }
          }
        },
        "security" : [ ]
      }
    },
    "/api/auth/refresh" : {
      "post" : {
        "tags" : [ "認証" ],
        "summary" : "トークン再発行",
        "description" : "`refresh_token` クッキーを検証し、アクセストークンとリフレッシュトークンの両方を\n**ローテーション**して再発行する（使い終わった古いリフレッシュトークンは失効する）。\n\nリクエストボディは不要。認証は `refresh_token` クッキーのみで行う。\nこのクッキーは `Path=/api/auth` に限定されているため、\n他のAPIパスへのリクエストには送信されない。\n\n典型的な使い方: 任意のAPIが401を返したら本APIを1回呼び、成功したら元のリクエストを再試行する。\n401が返った場合は再ログインが必要。\n",
        "operationId" : "refresh",
        "parameters" : [ {
          "name" : "refresh_token",
          "in" : "cookie",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "401" : {
            "description" : "リフレッシュトークンが無い・失効・再利用検知。再ログインが必要",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "INVALID_REFRESH_TOKEN",
                    "message" : "セッションの有効期限が切れました。再度ログインしてください"
                  }
                }
              }
            }
          },
          "429" : {
            "description" : "リクエストが多すぎる。Retry-Afterヘッダに再試行までの秒数が入る",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "RATE_LIMITED",
                    "message" : "アクセスが多すぎます。しばらく待ってからやり直してください"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "サーバー側の想定外エラー。詳細はクライアントへ返さずサーバーログに記録される",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "INTERNAL_ERROR",
                    "message" : "予期しないエラーが発生しました"
                  }
                }
              }
            }
          }
        },
        "security" : [ ]
      }
    },
    "/api/auth/logout" : {
      "post" : {
        "tags" : [ "認証" ],
        "summary" : "ログアウト",
        "description" : "リフレッシュトークンをサーバー側で失効させ、両クッキーを空・MaxAge=0で上書きして削除する。\n\n`refresh_token` クッキーが無い場合もエラーにせず200を返す（冪等）。\nログアウト処理がクライアント側の状態に依存して失敗しないようにするため。\n",
        "operationId" : "logout",
        "parameters" : [ {
          "name" : "refresh_token",
          "in" : "cookie",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "401" : {
            "description" : "未認証。auth_tokenクッキーが無い、または失効している",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "UNAUTHENTICATED",
                    "message" : "認証が必要です"
                  }
                }
              }
            }
          },
          "429" : {
            "description" : "リクエストが多すぎる。Retry-Afterヘッダに再試行までの秒数が入る",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "RATE_LIMITED",
                    "message" : "アクセスが多すぎます。しばらく待ってからやり直してください"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "サーバー側の想定外エラー。詳細はクライアントへ返さずサーバーログに記録される",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "INTERNAL_ERROR",
                    "message" : "予期しないエラーが発生しました"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login" : {
      "post" : {
        "tags" : [ "認証" ],
        "summary" : "ログイン",
        "description" : "成功時に2つのクッキーを発行する。\n\n- `auth_token` — アクセストークン(JWT・15分・`Path=/`)\n- `refresh_token` — リフレッシュトークン(opaque・7日・`Path=/api/auth`)\n\nどちらもhttpOnlyのためJavaScriptからは読めない。\nブラウザ以外のクライアントはSet-Cookieを保持して送り返す必要がある。\n",
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "資格情報が一致しない。メールとパスワードのどちらが誤りかは区別せず同じ応答を返す",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "INVALID_CREDENTIALS",
                    "message" : "メールアドレスまたはパスワードが正しくありません"
                  }
                }
              }
            }
          },
          "429" : {
            "description" : "試行が多すぎる。Retry-Afterヘッダに再試行までの秒数が入る。\n軸が2つあり、codeで区別できる。\n`TOO_MANY_LOGIN_ATTEMPTS` は同一アカウントへの失敗が上限に達した場合で、\n多数のIPから1つのアカウントを狙う形に効く。\n`RATE_LIMITED` は送信元IP単位の上限に達した場合。\nどちらも、未登録のアドレスかどうかで挙動を変えない\n(変えると「制限された = 実在する」という手掛かりを与えるため)。\n",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "TOO_MANY_LOGIN_ATTEMPTS",
                    "message" : "ログインの試行が多すぎます。しばらく待ってからやり直してください"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "リクエストが不正（バリデーション違反、型不一致、壊れたJSON）",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "VALIDATION_ERROR",
                    "message" : "本文は280文字以内で入力してください"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "サーバー側の想定外エラー。詳細はクライアントへ返さずサーバーログに記録される",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "INTERNAL_ERROR",
                    "message" : "予期しないエラーが発生しました"
                  }
                }
              }
            }
          }
        },
        "security" : [ ]
      }
    },
...
    "/api/auth/me" : {
      "get" : {
        "tags" : [ "認証" ],
        "summary" : "ログイン中ユーザーの取得",
        "description" : "`auth_token` クッキーから解決した現在のユーザーを返す。\n\nアプリ起動時のログイン状態の確認に使う。401が返ればログイン画面へ、\n200が返ればそのまま利用を継続してよい。\n",
        "operationId" : "me",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "未認証。auth_tokenクッキーが無い、または失効している",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "UNAUTHENTICATED",
                    "message" : "認証が必要です"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "サーバー側の想定外エラー。詳細はクライアントへ返さずサーバーログに記録される",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "error" : {
                    "code" : "INTERNAL_ERROR",
                    "message" : "予期しないエラーが発生しました"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "ApiResponseVoid" : {
        "type" : "object",
        "properties" : {
          "data" : { }
        },
        "required" : [ "data" ]
      },
...
      "SignupRequest" : {
        "type" : "object",
        "description" : "新規登録リクエスト",
        "properties" : {
          "email" : {
            "type" : "string",
            "format" : "email",
            "description" : "メールアドレス。既に登録済みの場合は400 `EMAIL_ALREADY_EXISTS`",
            "example" : "user@example.com",
            "minLength" : 1
          },
          "password" : {
            "type" : "string",
            "description" : "パスワード（8〜72文字）。上限72はBCryptが扱えるバイト長に由来する",
            "example" : "password1234",
            "maxLength" : 72,
            "minLength" : 8
          },
          "displayName" : {
            "type" : "string",
            "description" : "表示名（1〜100文字）。他ユーザーに表示される名前",
            "example" : "山田太郎",
            "maxLength" : 100,
            "minLength" : 0
          }
        },
        "required" : [ "displayName", "email", "password" ]
      },
      "ApiResponseUserResponse" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/UserResponse"
          }
        },
        "required" : [ "data" ]
      },
      "UserResponse" : {
        "type" : "object",
        "description" : "ログイン中ユーザー自身の情報。メールアドレスを含むため、**自分に関するAPIでのみ返る**\n（他人の情報は ProfileResponse / UserSummaryResponse で返り、メールアドレスは含まれない）",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "ユーザーID",
            "example" : 7
          },
          "email" : {
            "type" : "string",
            "description" : "メールアドレス。自分自身にのみ返る",
            "example" : "user@example.com"
          },
          "displayName" : {
            "type" : "string",
            "description" : "表示名（100文字以内）",
            "example" : "山田太郎"
          },
          "bio" : {
            "type" : [ "string", "null" ],
            "description" : "自己紹介（500文字以内）。未設定なら null",
            "example" : "バックエンドを担当しています"
          },
          "avatarUrl" : {
            "type" : [ "string", "null" ],
            "description" : "アイコンの署名付きURL（有効期限24時間）。未設定なら null"
          }
        },
        "required" : [ "avatarUrl", "bio", "displayName", "email", "id" ]
      },
      "LoginRequest" : {
        "type" : "object",
        "description" : "ログインリクエスト",
        "properties" : {
          "email" : {
            "type" : "string",
            "format" : "email",
            "description" : "登録済みのメールアドレス",
            "example" : "user@example.com",
            "minLength" : 1
          },
          "password" : {
            "type" : "string",
            "description" : "パスワード。メールアドレスとパスワードのどちらが誤っていても同じ401 `INVALID_CREDENTIALS` を返す（どちらが存在するかを推測させないため）",
            "example" : "password1234",
            "minLength" : 1
          }
        },
        "required" : [ "email", "password" ]
      },
...
      "ApiError" : {
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/ErrorBody"
          }
        },
        "required" : [ "error" ]
      },
      "ErrorBody" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "message" ]
      }
    },
    "securitySchemes" : {
      "cookieAuth" : {
        "type" : "apiKey",
        "description" : "httpOnlyクッキーに格納されたJWT。ログイン系APIがSet-Cookieで発行する。",
        "name" : "auth_token",
        "in" : "cookie"
      }
    }
  }
}